refactor(config): split validation god file
